Added Android 12 Classic Notfication and Code Refactoring
